users, authentication using JWT and Passport, guarded routes and logging

+++ src/recipes/recipes.service.ts
@@ -6,6 +6,7 @@ import { GetRecipesFilterDto } from './dto/get-recipes-filter-dto';
 import { RecipeRepository } from './recipe-repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Recipe } from './recipe.entity';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class RecipesService {
@@ -17,31 +18,45 @@ export class RecipesService {
 
     }
 
-    async getRecipes(filterDto: GetRecipesFilterDto): Promise<Recipe[]> {
-        return this.recipeRepository.getRecipes(filterDto);
+    async getRecipes(
+        filterDto: GetRecipesFilterDto,
+        user: User,
+    ): Promise<Recipe[]> {
+        return this.recipeRepository.getRecipes(filterDto, user);
     }
 
-    async getRecipeById(id: number): Promise<Recipe> {
-        const found = await this.recipeRepository.findOne(id);
+    async getRecipeById(
+        id: number, user: User): Promise<Recipe> {
+        const found = await this.recipeRepository.findOne({ where: { id, userId: user.id } });
         if (!found) {
             throw new NotFoundException(`Recipe with id "${id}" not found`);
         }
         return found;
     }
 
-    async createRecipe(createRecipeDto: CreateRecipeDto): Promise<Recipe> {
-        return this.recipeRepository.createRecipe(createRecipeDto);
+    async createRecipe(
+        createRecipeDto: CreateRecipeDto,
+        user: User,
+    ): Promise<Recipe> {
+        return this.recipeRepository.createRecipe(createRecipeDto, user);
     }
 
-    async deleteRecipe(id: number): Promise<void> {
-        const result = await this.recipeRepository.delete(id);
+    async deleteRecipe(
+        id: number,
+        user: User,
+    ): Promise<void> {
+        const result = await this.recipeRepository.delete({id, userId: user.id});
         if (result.affected === 0) {
             throw new NotFoundException(`Recipe with id "${id}" not found`);
         }
     }
 
-    async updateRecipeStatus(id: number, status: RecipeStatus): Promise<Recipe> {
-        const recipe = await this.getRecipeById(id);
+    async updateRecipeStatus(
+        id: number,
+        status: RecipeStatus,
+        user: User,
+    ): Promise<Recipe> {
+        const recipe = await this.getRecipeById(id, user);
         recipe.status = status;
         await recipe.save();
         return recipe;
